Fix backend payment templates not being found
Fix backend payment templates not being found. Might want to make it a config setting. But this works fine now - and as PDF invoices are always created in the adminhtml env maybe this change is for the better. Just tested on our production and seems to solve all errors

Fixes #16